+++ tests/bootstrap.php
@@ -1,0 +1,8 @@
+<?php
+
+$loader = require 'vendor/autoload.php';
+
+require 'DatatablesTestCase.php';
+
+// Add stuff to autoload
+$loader->add('', 'tests/laravel');
